allow enabling ArgoCD core workloads independently

+++ controllers/argocd/statefulset.go
@@ -391,8 +391,8 @@ func (r *ReconcileArgoCD) reconcileRedisStatefulSet(cr *argoproj.ArgoCD) error {
 
 	existing := newStatefulSetWithSuffix("redis-ha-server", "redis", cr)
 	if argoutil.IsObjectFound(r.Client, cr.Namespace, existing.Name, existing) {
-		if !cr.Spec.HA.Enabled {
-			// StatefulSet exists but HA enabled flag has been set to false, delete the StatefulSet
+		if !(cr.Spec.HA.Enabled && IsComponentEnabled(cr.Spec.Redis.Enabled)) {
+			// StatefulSet exists but either HA or component enabled flag has been set to false, delete the StatefulSet
 			return r.Client.Delete(context.TODO(), existing)
 		}
 
@@ -422,6 +422,11 @@ func (r *ReconcileArgoCD) reconcileRedisStatefulSet(cr *argoproj.ArgoCD) error {
 		}
 
 		return nil // StatefulSet found, do nothing
+	}
+
+	if !IsComponentEnabled(cr.Spec.Redis.Enabled) {
+		log.Info("Redis disabled. Skipping starting Redis.") // Redis not enabled, do nothing.
+		return nil
 	}
 
 	if !cr.Spec.HA.Enabled {
@@ -649,6 +654,11 @@ func (r *ReconcileArgoCD) reconcileApplicationControllerStatefulSet(cr *argoproj
 
 	existing := newStatefulSetWithSuffix("application-controller", "application-controller", cr)
 	if argoutil.IsObjectFound(r.Client, cr.Namespace, existing.Name, existing) {
+		if !IsComponentEnabled(cr.Spec.Controller.Enabled) {
+			log.Info("Existing application controller found but should be disabled. Deleting Application Controller")
+			// Delete existing deployment for Application Controller, if any ..
+			return r.Client.Delete(context.TODO(), existing)
+		}
 		actualImage := existing.Spec.Template.Spec.Containers[0].Image
 		desiredImage := getArgoContainerImage(cr)
 		changed := false
@@ -694,6 +704,11 @@ func (r *ReconcileArgoCD) reconcileApplicationControllerStatefulSet(cr *argoproj
 			return r.Client.Update(context.TODO(), existing)
 		}
 		return nil // StatefulSet found with nothing to do, move along...
+	}
+
+	if !IsComponentEnabled(cr.Spec.Controller.Enabled) {
+		log.Info("Application Controller disabled. Skipping starting application controller.")
+		return nil
 	}
 
 	// Delete existing deployment for Application Controller, if any ..
